añadiendo mensajes a alerts

+++ src/front/js/pages/asociacion.js
@@ -1,9 +1,9 @@
-
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 
 export default function Asociacion() {
     const [animals, setAnimals] = useState(null);
+    const [successMessage, setSuccessMessage] = useState(false);
 
     const token = localStorage.getItem("token");
 
@@ -21,16 +21,33 @@ export default function Asociacion() {
     };
 
     const deleteAnimal = async (animal) => {
-        const response = await fetch(process.env.BACKEND_URL + "/api/animal/" + animal.id, {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json",
-                Authorization: `Bearer ${token}`,
-            },
-        });
-        const data = await response.json();
-        console.log(data);
-        fetchAnimal();
+        const confirmDelete = window.confirm("¿Estás seguro de eliminar este animal?");
+
+        if (confirmDelete) {
+            try {
+                const response = await fetch(process.env.BACKEND_URL + "/api/animal/" + animal.id, {
+                    method: "DELETE",
+                    headers: {
+                        "Content-Type": "application/json",
+                        Authorization: `Bearer ${token}`,
+                    },
+                });
+
+                if (!response.ok) {
+                    throw new Error("Error al eliminar el animal");
+                }
+
+                const data = await response.json();
+                console.log(data);
+                fetchAnimal();
+                setSuccessMessage(true);
+                setTimeout(() => {
+                    setSuccessMessage(false);
+                }, 3000);
+            } catch (error) {
+                console.error(error);
+            }
+        }
     };
 
     useEffect(() => {
@@ -39,6 +56,11 @@ export default function Asociacion() {
 
     return (
         <>
+            {successMessage && (
+                <div className="alert alert-success" role="alert">
+                    Animal eliminado correctamente
+                </div>
+            )}
             <div>
                 <h1>Recuperando todos los animales de la Asociación que hizo login</h1>
                 <Link to="/animalForm">
